+++ cms-server/src/main/java/com/condation/cms/server/handler/auth/JettyAuthenticationHandler.java
@@ -21,8 +21,6 @@ package com.condation.cms.server.handler.auth;
  * <http://www.gnu.org/licenses/gpl-3.0.html>.
  * #L%
  */
-
-
 import com.condation.cms.api.cache.ICache;
 import com.condation.cms.api.feature.features.AuthFeature;
 import com.condation.cms.api.request.RequestContext;
@@ -60,10 +58,6 @@ public class JettyAuthenticationHandler extends Handler.Abstract {
 	private final ICache<String, AtomicInteger> loginFails;
 
 //	LoadingCache<String, AtomicInteger> loginFails = Caffeine.newBuilder()
-//			.maximumSize(10_000)
-//			.expireAfterWrite(Duration.ofMinutes(1))
-//			.expireAfterAccess(Duration.ofMinutes(1))
-//			.build(key -> new AtomicInteger(0));
 
 	static final int ATTEMPTS_TO_BLOCK = 3;
 
@@ -112,11 +106,11 @@ public class JettyAuthenticationHandler extends Handler.Abstract {
 							}
 
 							if (authPath.allowed(userOpt.get())) {
-								
+
 								var requestContext = (RequestContext) request.getAttribute(CreateRequestContextFilter.REQUEST_CONTEXT);
 								requestContext.add(AuthFeature.class, new AuthFeature(username));
-								
-								loginFails.invalidate(clientAddress(request));
+
+								loginFails.invalidate(RequestUtil.clientAddress(request));
 								return false;
 							}
 
@@ -143,13 +137,8 @@ public class JettyAuthenticationHandler extends Handler.Abstract {
 		callback.succeeded();
 	}
 
-	private String clientAddress(Request request) {
-		return ((InetSocketAddress) request.getConnectionMetaData().getRemoteSocketAddress())
-				.getAddress().getHostAddress();
-	}
-
 	private AtomicInteger getClientLoginCounter(Request request) {
-		return loginFails.get(clientAddress(request));
+		return loginFails.get(RequestUtil.clientAddress(request));
 	}
 
 }