+++ test-server/themes/demo/extensions/theme.extension.js
@@ -8,3 +8,26 @@ $hooks.registerAction("system/content/shortcodes", (context) => {
 	)
 	return null;
 })
+
+$hooks.registerAction("system/template/component", (context) => {
+	context.arguments().get("components").put(
+			"component",
+			(params) => `<div style="color: ${params.color}">${params.message}</div>`
+	)
+	return null;
+})
+
+$hooks.registerFilter("module/ui/translations", (context) => {
+	var translations = context.value()
+	
+	translations.en["field.title"] = "Title";
+	translations.de["field.title"] = "Titel";
+	
+	translations.en["field.parent.text"] = "Parent-Text";
+	translations.de["field.parent.text"] = "Eltern-Text";
+	
+	translations.en["field.description"] = "Description";
+	translations.de["field.description"] = "Beschreibung";
+	
+	return translations;
+})