+++ CHANGELOG.md
@@ -6,7 +6,9 @@ see wiki for more information: [wiki](https://github.com/thmarx/cms/wiki)
 
 # changelog
 
-## 7.9.0
+## 8.0.0
+
+* **BREAKING CHANGE** Sorted sections now use the _layout.order_ meta attribute for sorting
 
 * **BUGFIX** TemplateEngine should use cache only if activated [456](https://github.com/CondationCMS/cms-server/issues/456)
 * **MAINTENANCE** multiple dependencies updated
@@ -20,6 +22,8 @@ see wiki for more information: [wiki](https://github.com/thmarx/cms/wiki)
 * **FEATURE** Switch password has to secure algorithm [472](https://github.com/CondationCMS/cms-server/issues/472)
 * **FEATURE** Simple http api for basic use cases [479](https://github.com/CondationCMS/cms-server/issues/479)
 * **FEATURE** Add spa mode for sites [476](https://github.com/CondationCMS/cms-server/issues/476)
+* **FEATURE** Custom repository urls for modules, extensions and themes [466](https://github.com/CondationCMS/cms-server/issues/466)
+* **FEATURE** UI to manage content [PR-446](https://github.com/CondationCMS/cms-server/pull/446)
 
 ### Developer experience